requirements.txt
matplotlib
liquidcrystal-i2c-linux
